Migrate to SAF (Storage Access Framework)

+++ app/src/main/java/awais/instagrabber/services/DeleteImageIntentService.java
@@ -4,13 +4,14 @@ import android.app.IntentService;
 import android.app.PendingIntent;
 import android.content.Context;
 import android.content.Intent;
+import android.net.Uri;
 import android.util.Log;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.core.app.NotificationManagerCompat;
+import androidx.documentfile.provider.DocumentFile;
 
-import java.io.File;
 import java.util.Random;
 
 import awais.instagrabber.utils.TextUtils;
@@ -39,7 +40,10 @@ public class DeleteImageIntentService extends IntentService {
         if (intent != null && Intent.ACTION_DELETE.equals(intent.getAction()) && intent.hasExtra(EXTRA_IMAGE_PATH)) {
             final String path = intent.getStringExtra(EXTRA_IMAGE_PATH);
             if (TextUtils.isEmpty(path)) return;
-            final File file = new File(path);
+            // final File file = new File(path);
+            final Uri parse = Uri.parse(path);
+            if (parse == null) return;
+            final DocumentFile file = DocumentFile.fromSingleUri(getApplicationContext(), parse);
             boolean deleted;
             if (file.exists()) {
                 deleted = file.delete();
@@ -58,11 +62,11 @@ public class DeleteImageIntentService extends IntentService {
 
     @NonNull
     public static PendingIntent pendingIntent(@NonNull final Context context,
-                                              @NonNull final String imagePath,
+                                              @NonNull final DocumentFile imagePath,
                                               final int notificationId) {
         final Intent intent = new Intent(context, DeleteImageIntentService.class);
         intent.setAction(Intent.ACTION_DELETE);
-        intent.putExtra(EXTRA_IMAGE_PATH, imagePath);
+        intent.putExtra(EXTRA_IMAGE_PATH, imagePath.getUri().toString());
         intent.putExtra(EXTRA_NOTIFICATION_ID, notificationId);
         return PendingIntent.getService(context, random.nextInt(), intent, PendingIntent.FLAG_UPDATE_CURRENT);
     }
